Support for both long & short living caches
