+++ packages/smooth_app/test/utils/path_provider_mock.dart
@@ -1,0 +1,51 @@
+import 'dart:io';
+
+import 'package:mockito/mockito.dart';
+import 'package:path_provider_platform_interface/path_provider_platform_interface.dart';
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+
+class MockedPathProviderPlatform extends Mock
+    with MockPlatformInterfaceMixin
+    implements PathProviderPlatform {
+  @override
+  Future<String?> getTemporaryPath() async {
+    return Directory.systemTemp.path;
+  }
+
+  @override
+  Future<String?> getApplicationSupportPath() async {
+    return Directory.systemTemp.path;
+  }
+
+  @override
+  Future<String?> getLibraryPath() async {
+    return Directory.systemTemp.path;
+  }
+
+  @override
+  Future<String?> getApplicationDocumentsPath() async {
+    return Directory.systemTemp.path;
+  }
+
+  @override
+  Future<String?> getExternalStoragePath() async {
+    return Directory.systemTemp.path;
+  }
+
+  @override
+  Future<List<String>?> getExternalCachePaths() async {
+    return <String>[Directory.systemTemp.path];
+  }
+
+  @override
+  Future<List<String>?> getExternalStoragePaths({
+    StorageDirectory? type,
+  }) async {
+    return <String>[Directory.systemTemp.path];
+  }
+
+  @override
+  Future<String?> getDownloadsPath() async {
+    return Directory.systemTemp.path;
+  }
+}
